# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/avatar/": {
      "post": {
        "tags": [
          "Standalone Avatar"
        ],
        "summary": "Create Standalone Avatar",
        "description": "Create a new avatar independently (before story creation).",
        "operationId": "create_standalone_avatar_avatar__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/avatar/{avatar_id}": {
      "get": {
        "tags": [
          "Standalone Avatar"
        ],
        "summary": "Get Standalone Avatar",
        "description": "Retrieve an avatar by ID.",
        "operationId": "get_standalone_avatar_avatar__avatar_id__get",
        "parameters": [
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Avatar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/story/{story_id}/avatar/": {
      "post": {
        "tags": [
          "Avatar"
        ],
        "summary": "Create Avatar",
        "description": "Create a new avatar for the story.",
        "operationId": "create_avatar_story__story_id__avatar__post",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/story/{story_id}/avatar/{avatar_id}": {
      "get": {
        "tags": [
          "Avatar"
        ],
        "summary": "Get Avatar",
        "description": "Retrieve a single avatar by ID for the story.",
        "operationId": "get_avatar_story__story_id__avatar__avatar_id__get",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Avatar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Avatar"
        ],
        "summary": "Update Avatar",
        "description": "Update an existing avatar for the story.",
        "operationId": "update_avatar_story__story_id__avatar__avatar_id__put",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Avatar Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/story/{story_id}/message/{message_id}": {
      "put": {
        "tags": [
          "Message"
        ],
        "summary": "Update Message",
        "description": "Update a message.",
        "operationId": "update_message_story__story_id__message__message_id__put",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Message"
        ],
        "summary": "Delete Message",
        "description": "Delete a message.",
        "operationId": "delete_message_story__story_id__message__message_id__delete",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AvatarCreate": {
        "properties": {
          "avatar_name": {
            "type": "string",
            "title": "Avatar Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "avatar_name"
        ],
        "title": "AvatarCreate",
        "description": "Creating new avatar character.\nUsed during character creation phase before the story begins.\n\nUser provides avatar name and description; System generates avatar_id and timestamps automatically."
      },
      "AvatarRead": {
        "properties": {
          "avatar_id": {
            "type": "integer",
            "title": "Avatar Id"
          },
          "avatar_name": {
            "type": "string",
            "title": "Avatar Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "avatar_id",
          "avatar_name",
          "created_at"
        ],
        "title": "AvatarRead",
        "description": "Returning avatar information via API.\nUsed when displaying avatar list/avatar details to users."
      },
      "AvatarUpdate": {
        "properties": {
          "avatar_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "AvatarUpdate",
        "description": "Updating avatar information.\nUsed when user wants to rename their avatar or update description.\n\nOptional."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageRead": {
        "properties": {
          "message_id": {
            "type": "integer",
            "title": "Message Id"
          },
          "story_id": {
            "type": "integer",
            "title": "Story Id"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "message_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Label"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "story_id",
          "content",
          "role",
          "created_at"
        ],
        "title": "MessageRead",
        "description": "Returning message information.\nUsed when displaying conversation history/story transcript."
      },
      "MessageUpdate": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "message_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Label"
          }
        },
        "type": "object",
        "title": "MessageUpdate",
        "description": "Updating the message content when needed.\nUsed when user edits their previous message in conversation.\n\nOptional."
      },
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "system",
          "assistant",
          "internal",
          "metadata"
        ],
        "title": "Role",
        "description": "Enum for message source types in story conversation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}